学期终结，HD两个发射机构都可自动发射了，相关指令和参数的存储未完善

+++ USER/main.h
@@ -18,7 +18,7 @@ float pur_pitch,pur_roll;
 bool roll_flag=false,pitch_flag=false;
 
 struct {
-	int d,dd,d_last,now;
+	float d,dd,d_last,now;
 	int cnt;
 }pitch,roll;
